+++ test/AMM/CP.t.sol
@@ -7,31 +7,30 @@ import "@hack/amm/CP.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@hack/myCoin/MyERC20.sol";
 
-
-contract CPTest is Test{
+contract CPTest is Test {
     CP public cp;
     MyERC20 public t1;
     MyERC20 public t2;
 
-    function setUp() public{
+    function setUp() public {
         t1 = new MyERC20("token1", "t1");
         t2 = new MyERC20("token2", "t2");
-        cp = new CP(address (t1), address (t2));
+        cp = new CP(address(t1), address(t2));
     }
-    
-    function testSwap() public{
+
+    function testSwap() public {
         address user = vm.addr(1);
         vm.startPrank(user);
-        t1.mint(address(user),200);
-        t2.mint(address(user),200);
-        t1.approve(address(cp),50);
-        t2.approve(address(cp),90);
+        t1.mint(address(user), 200);
+        t2.mint(address(user), 200);
+        t1.approve(address(cp), 50);
+        t2.approve(address(cp), 90);
         cp.addLiquidity(50, 90);
         console.log(address(t1));
-        uint initial2 = t2.balanceOf(address(cp));
-        uint initial1 = t1.balanceOf(address(cp));
+        uint256 initial2 = t2.balanceOf(address(cp));
+        uint256 initial1 = t1.balanceOf(address(cp));
         console.log(t1.balanceOf(address(cp)));
-        t1.approve(address(cp),20);
+        t1.approve(address(cp), 20);
 
         cp.swap(address(t1), 20);
         console.log(t1.balanceOf(address(cp)));
